+++ apis/routing/src/index.ts
@@ -10,111 +10,16 @@
  * Learn more at https://developers.cloudflare.com/workers/
  */
 
+import { ChainId, Currency, JSBI } from '@pancakeswap/sdk'
+import { SmartRouter, StablePool, V2Pool, V3Pool } from '@pancakeswap/smart-router/evm'
 import { CORS_ALLOW, handleCors, wrapCorsHeader } from '@pancakeswap/worker-utils'
+import { GraphQLClient } from 'graphql-request'
 import { Router } from 'itty-router'
 import { error, json, missing } from 'itty-router-extras'
-import { z } from 'zod'
-import { ChainId, Currency, JSBI, TradeType } from '@pancakeswap/sdk'
-import { PoolType, SmartRouter, StablePool, V2Pool, V3Pool } from '@pancakeswap/smart-router/evm'
-import { FeeAmount } from '@pancakeswap/v3-sdk'
-import { GraphQLClient } from 'graphql-request'
-import { viemProviders } from './provider'
 import { sendLog } from './log'
+import { viemProviders } from './provider'
 
 const { parseCurrency, parseCurrencyAmount, parsePool, serializeTrade } = SmartRouter.Transformer
-
-const zChainId = z.nativeEnum(ChainId)
-const zFee = z.nativeEnum(FeeAmount)
-const zTradeType = z.nativeEnum(TradeType)
-const zPoolType = z.nativeEnum(PoolType)
-const zPoolTypes = z.array(zPoolType)
-const zAddress = z.string().regex(/^0x[a-fA-F0-9]{40}$/)
-const zBigNumber = z.string().regex(/^[0-9]+$/)
-const zCurrency = z
-  .object({
-    address: zAddress,
-    decimals: z.number(),
-    symbol: z.string(),
-  })
-  .required()
-const zCurrencyAmount = z
-  .object({
-    currency: zCurrency,
-    value: zBigNumber,
-  })
-  .required()
-
-const zV2Pool = z
-  .object({
-    type: zPoolType,
-    reserve0: zCurrencyAmount,
-    reserve1: zCurrencyAmount,
-  })
-  .required()
-const zV3Pool = z
-  .object({
-    type: zPoolType,
-    token0: zCurrency,
-    token1: zCurrency,
-    fee: zFee,
-    liquidity: zBigNumber,
-    sqrtRatioX96: zBigNumber,
-    tick: z.number(),
-    address: zAddress,
-    token0ProtocolFee: z.string(),
-    token1ProtocolFee: z.string(),
-  })
-  .required()
-const zStablePool = z
-  .object({
-    type: zPoolType,
-    balances: z.array(zCurrencyAmount),
-    amplifier: zBigNumber,
-    fee: z.string(),
-  })
-  .required()
-const zPools = z.array(z.union([zV2Pool, zV3Pool, zStablePool]))
-
-const zGetParams = z
-  .object({
-    chainId: zChainId,
-    tradeType: zTradeType,
-    amount: zCurrencyAmount,
-    currency: zCurrency,
-    gasPriceWei: zBigNumber.optional(),
-    maxHops: z.number().optional(),
-    maxSplits: z.number().optional(),
-    blockNumber: zBigNumber.optional(),
-    poolTypes: zPoolTypes.optional(),
-  })
-  .required({
-    chainId: true,
-    tradeType: true,
-    amount: true,
-    currency: true,
-    candidatePools: true,
-  })
-
-const zPostParams = z
-  .object({
-    chainId: zChainId,
-    tradeType: zTradeType,
-    amount: zCurrencyAmount,
-    currency: zCurrency,
-    candidatePools: zPools,
-    gasPriceWei: zBigNumber.optional(),
-    maxHops: z.number().optional(),
-    maxSplits: z.number().optional(),
-    blockNumber: zBigNumber.optional(),
-    poolTypes: zPoolTypes.optional(),
-  })
-  .required({
-    chainId: true,
-    tradeType: true,
-    amount: true,
-    currency: true,
-    candidatePools: true,
-  })
 
 const router = Router()
 
@@ -173,7 +78,7 @@ async function sha256(message: string) {
 }
 
 router.get('/v0/quote', async (req, event: FetchEvent) => {
-  const parsed = zGetParams.safeParse(req.query)
+  const parsed = SmartRouter.APISchema.zRouterGetParams.safeParse(req.query)
   if (parsed.success === false) {
     return error(400, 'Invalid params')
   }
@@ -239,7 +144,7 @@ function timeout(seconds: number) {
 
 router.post('/v0/quote', async (req, event) => {
   const body = (await req.json?.()) as any
-  const parsed = zPostParams.safeParse(body)
+  const parsed = SmartRouter.APISchema.zRouterPostParams.safeParse(body)
   if (parsed.success === false) {
     return error(400, 'Invalid params')
   }
