+++ packages/smart-router/evm/v3-router/smartRouter.ts
@@ -28,3 +28,4 @@ export {
   Transformer,
 } from './utils'
 export { getPairCombinations } from './functions'
+export * as APISchema from './schema'
